+++ src/ValueObjects/NormalTransaction.php
@@ -2,7 +2,7 @@
 
 namespace Nemo\Etherscan\ValueObjects;
 
-use Brick\Math\BigInteger;
+use Brick\Math\BigDecimal;
 use Brick\Math\Exception\MathException;
 use Carbon\CarbonImmutable;
 use Nemo\Etherscan\Support\Numbers;
@@ -10,7 +10,7 @@ use Nemo\Etherscan\Support\Numbers;
 class NormalTransaction
 {
     public function __construct(
-        public BigInteger $blockNumber,
+        public BigDecimal $blockNumber,
         public CarbonImmutable $timeStamp,
         public string $hash,
         public int $nonce,
@@ -18,16 +18,16 @@ class NormalTransaction
         public int $transactionIndex,
         public string $from,
         public string $to,
-        public BigInteger $value,
-        public BigInteger $gas,
-        public BigInteger $gasPrice,
+        public BigDecimal $value,
+        public BigDecimal $gas,
+        public BigDecimal $gasPrice,
         public bool $isError,
         public string $receiptStatus,
         public string $input,
         public string $contractAddress,
-        public BigInteger $cumulativeGasUsed,
-        public BigInteger $gasUsed,
-        public BigInteger $confirmations,
+        public BigDecimal $cumulativeGasUsed,
+        public BigDecimal $gasUsed,
+        public BigDecimal $confirmations,
         public ?string $methodId,
         public ?string $functionName,
     ) {
@@ -40,7 +40,7 @@ class NormalTransaction
     public static function fromResponse(array $response): NormalTransaction
     {
         return new NormalTransaction(
-            blockNumber: BigInteger::of($response['blockNumber']),
+            blockNumber: BigDecimal::of($response['blockNumber']),
             timeStamp: CarbonImmutable::createFromTimestamp($response['timeStamp']),
             hash: $response['hash'],
             nonce: (int) $response['nonce'],
@@ -48,16 +48,16 @@ class NormalTransaction
             transactionIndex: (int) $response['transactionIndex'],
             from: $response['from'],
             to: $response['to'],
-            value: BigInteger::of($response['value']),
-            gas: BigInteger::of($response['gas']),
-            gasPrice: BigInteger::of($response['gasPrice']),
+            value: BigDecimal::of($response['value']),
+            gas: BigDecimal::of($response['gas']),
+            gasPrice: BigDecimal::of($response['gasPrice']),
             isError: (bool) $response['isError'],
             receiptStatus: $response['txreceipt_status'],
             input: $response['input'],
             contractAddress: $response['contractAddress'],
-            cumulativeGasUsed: BigInteger::of($response['cumulativeGasUsed']),
-            gasUsed: BigInteger::of($response['gasUsed']),
-            confirmations: BigInteger::of($response['confirmations']),
+            cumulativeGasUsed: BigDecimal::of($response['cumulativeGasUsed']),
+            gasUsed: BigDecimal::of($response['gasUsed']),
+            confirmations: BigDecimal::of($response['confirmations']),
             methodId: $response['methodId'] ?? null,
             functionName: $response['functionName'] ?? null,
         );
@@ -65,6 +65,6 @@ class NormalTransaction
 
     public function resolveValue(): float
     {
-        return $this->value->isGreaterThan(BigInteger::zero()) ? Numbers::scaleToFloat($this->value, 18) : 0;
+        return $this->value->isGreaterThan(BigDecimal::zero()) ? Numbers::scaleToFloat($this->value, 18) : 0;
     }
 }
